correct color in semantickitti_label and semanticfusionportable_label

+++ nvblox/include/nvblox/core/voxels.h
@@ -62,7 +62,7 @@ typedef Eigen::Matrix<float, kTotalNumberOfLabels, kTotalNumberOfLabels>
 struct SemanticVoxel {
   // Initialize voxel to unknown label.
   // uint16_t semantic_label = 0u;
-  short int semantic_label = 0;
+  uint8_t semantic_label = 0;
   // Initialize voxel to uniform probability.
   // Use log odds! So uniform ditribution of 1/kTotalNumberOfLabels,
   // should be std::log(1/kTotalNumberOfLabels)
